+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 == 0.0.5 (unreleased)
 - options `print_to_stderr`, `print_recurring`
+- fix redis deprecated `pipelined` block arity (support for redis 5)
 
 == 0.0.4
 - added first_timestamp to deprecations (unix timestamp of first occurrence, not accurate because a worker with later timestamp may dump its deprecations earlier)